refactor: move nls code one level down
