+++ src/api/controllers/v1/nls/workflow.go
@@ -26,7 +26,7 @@ package controllers_v1
 import (
 	"fmt"
 
-	"github.com/Cray-HPE/cray-nls/src/api/models"
+	models "github.com/Cray-HPE/cray-nls/src/api/models/nls"
 	"github.com/Cray-HPE/cray-nls/src/api/services"
 	"github.com/Cray-HPE/cray-nls/src/utils"
 	"github.com/gin-gonic/gin"
